malloc free

+++ 0x0C-more_malloc_free/101-mul.c
@@ -10,8 +10,8 @@
 int main(int argc, char *argv[])
 {
 	unsigned long mul;
-	unsigned long i;
-	unsigned long j;
+	int i;
+	int j;
 
 	if (argc != 3)
 	{
